diffuir_restoration/requirements.txt
# DiffUIR Fingerprint Restoration Requirements

# Core deep learning frameworks
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Image processing and computer vision
opencv-python>=4.6.0
pillow>=9.0.0
scikit-image>=0.19.0
imageio>=2.19.0

# Numerical computing
numpy>=1.21.0
scipy>=1.8.0

# Evaluation metrics and visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.64.0
wandb>=0.13.0  # For experiment tracking
tensorboard>=2.10.0

# Data handling
pandas>=1.4.0
h5py>=3.7.0

# Model utilities
timm>=0.6.0  # For pre-trained models
einops>=0.4.0  # For tensor operations

# Diffusion model dependencies
diffusers>=0.10.0
transformers>=4.21.0
accelerate>=0.12.0

# Optional: GPU acceleration
# cuda-toolkit  # Uncomment if using CUDA

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
